+++ src/helper.cpp
@@ -8,7 +8,7 @@
 // Printing functions
 const Sys& Sys::p(char* c) const { std::cout << c; return *this; }
 const Sys& Sys::p(bool c) const { std::cout << c; return *this; }
-const Sys& Sys::p(float c) const { std::cout << c; return *this; }  
+const Sys& Sys::p(double c) const { std::cout << c; return *this; }  
 const Sys& Sys::p(int i) const { std::cout << i;  return *this; }
 const Sys& Sys::p(size_t i) const { std::cout << i;  return *this; }
 const Sys& Sys::p(const char* c) const { std::cout << c;  return *this; }
@@ -18,7 +18,7 @@ const Sys& Sys::pln(int i) const { std::cout << i << std::endl;  return *this; }
 const Sys& Sys::pln(char* c) const { std::cout << c << std::endl;  return *this; }
 const Sys& Sys::pln(bool c) const { std::cout << c << std::endl;  return *this; }  
 const Sys& Sys::pln(char c) const { std::cout << c << std::endl;  return *this; }
-const Sys& Sys::pln(float x) const { std::cout << x << std::endl;  return *this; }
+const Sys& Sys::pln(double x) const { std::cout << x << std::endl;  return *this; }
 const Sys& Sys::pln(size_t x) const { std::cout << x << std::endl;  return *this; }
 const Sys& Sys::pln(const char* c) const { std::cout << c << std::endl;  return *this; }
